Added license for werkzeug code
